more fixes

+++ src/admin/pages/GeneralSettings/PublicPagesPage.tsx
@@ -11,13 +11,10 @@ const PublicPagesPage = (): ReactElement => {
   const navigate = useNavigate();
   const { data: { pages = null } = {}, refetch } = useGetPagesQuery();
   const [deletePage] = useDeletePageMutation({ onCompleted: () => refetch() });
-
   const handleUpdatePage = (id: string) => () => navigate(id);
   const handleDeletePage = (id: string) => () => {
     // eslint-disable-next-line no-restricted-globals
-    if (confirm('Möchten Sie dies löschen?')) {
-      deletePage({ variables: { id } });
-    }
+    if (confirm('Möchten Sie dies löschen?')) deletePage({ variables: { id } });
   };
 
   return (
